requirements.txt
fastapi
uvicorn
pydantic
google-ads
